requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
sqlalchemy
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
pandas
openpyxl
reportlab
fastapi-mail
